+++ desktop/frontend/src/components/Composer.tsx
@@ -7,6 +7,7 @@ import { DedupIndex, sha256 } from "../lib/attachDedup";
 import { app, onFilesDropped } from "../lib/bridge";
 import { canUsePromptHistory, composerEnterAction, composerEscapeAction, composerMenuKeyAction, insertComposerNewline, isFnKeyEvent, isImeKeyEvent, promptHistoryDirectionFromEvent } from "../lib/composerKeyboard";
 import { cacheGeneration, loadOlder } from "../lib/composerHistory";
+import { sessionTurnsLabel } from "../lib/sessionCatalogPresentation";
 import { SPINNER_WORDS, useI18n, type Translator } from "../lib/i18n";
 import { detectShortcutPlatform, formatShortcutCombo, isReservedComposerHistoryShortcut, matchesShortcut, useShortcutComboLabel } from "../lib/keyboardShortcuts";
 import { fallbackCopyText } from "../lib/clipboard";
@@ -4119,23 +4120,19 @@ export function Composer({
                   </div>
                 ) : (
                   filteredPastChats.map((session, i) => {
-                    // PR-C2: hover preview uses only the SessionMeta fields we
-                    // already have on hand — no extra PreviewSession call, no
-                    // backend round-trip, no read of the full transcript.
-                    const turnsKnown = session.turnsState !== "unknown" && typeof session.turns === "number";
-                    const turnsPending = session.turnsState === "unknown";
+                    // Hover preview stays on SessionMeta and never reads the transcript.
+                    const turnsLabel = sessionTurnsLabel(session, t);
                     const ts = session.lastActivityAt || session.modTime || session.createdAt;
                     const preview = truncatePreview(session.preview);
                     const pathText = session.workspaceRoot || session.path;
                     const tooltipLabel =
-                      turnsKnown || turnsPending || ts || preview || pathText ? (
+                      turnsLabel || ts || preview || pathText ? (
                         <div className="past-chat-hover">
                           <div className="past-chat-hover__title">{pastChatTitle(session)}</div>
                           {preview && <div className="past-chat-hover__preview">{preview}</div>}
-                          {(turnsKnown || turnsPending || ts) && (
+                          {(turnsLabel || ts) && (
                             <div className="past-chat-hover__meta">
-                              {turnsPending && <span>{t("history.indexing")}</span>}
-                              {turnsKnown && <span>{t("composer.sessionTurns", { n: session.turns })}</span>}
+                              {turnsLabel && <span>{turnsLabel}</span>}
                               {ts && <span>· {fmtSessionTime(ts)}</span>}
                             </div>
                           )}
@@ -4155,11 +4152,7 @@ export function Composer({
                           <MessageSquare size={13} className="filemenu__icon" />
                           <span className="slashmenu__name slashmenu__name--file">
                             {pastChatTitle(session)}
-                            {turnsPending
-                              ? ` (${t("history.indexing")})`
-                              : turnsKnown
-                                ? ` (${t("composer.sessionTurns", { n: session.turns })})`
-                                : ""}
+                            {turnsLabel ? ` (${turnsLabel})` : ""}
                           </span>
                         </button>
                       </Tooltip>
@@ -4324,11 +4317,7 @@ export function Composer({
                   <MessageSquare size={15} />
                   <span>
                     {ref.title}
-                    {ref.turnsState === "unknown"
-                      ? ` (${t("history.indexing")})`
-                      : typeof ref.turns === "number"
-                        ? ` (${t("composer.sessionTurns", { n: ref.turns })})`
-                        : ""}
+                    {sessionTurnsLabel(ref, t) ? ` (${sessionTurnsLabel(ref, t)})` : ""}
                   </span>
                 </span>
               </Tooltip>
