+++ desktop/frontend/src/lib/bridge.ts
@@ -1,7 +1,4 @@
-// bridge is the seam between React and the Go kernel. The Wails shell calls bound
-// App methods and subscribes to runtime events; in a plain browser (`pnpm dev`),
-// a mock streams a canned turn through the same contract so the whole UI can be
-// developed and laid out without rebuilding the Go side.
+// Wails and the browser mock share this React-to-Go contract.
 
 // @ts-ignore `wails generate module` creates this locally; fresh checkouts keep
 // typecheck green by falling back to a disabled drift check below.
@@ -9,8 +6,8 @@ import type * as GeneratedApp from "../../wailsjs/go/main/App";
 import type { InvocationRequest } from "./invocationDisplay";
 
 import { addBreadcrumb } from "./breadcrumbs";
-import { asArray } from "./array";
 import { maybeShare } from "./queryCoalesce";
+import { makeMockSessionCatalogBindings } from "./sessionCatalogBridge";
 import { t } from "./i18n";
 import { providerIsConfigured, providerRequiresKey, removeProviderAccessesForMock } from "./providerModels";
 import { DEFAULT_STATUS_BAR_ITEMS, normalizeStatusBarItems } from "./statusBarItems";
@@ -87,12 +84,7 @@ import type {
   PluginInstallOptions,
   PluginView,
   ProjectNode,
-  ProjectTreeSnapshot,
-  ProjectTopicPageRequest,
-  ProjectTopicPage,
-  ProjectTopicKey,
-  SessionCatalogStatus,
-  ProjectTreeChangedV2,
+  SessionCatalogBindings,
   PromptHistoryEntry,
   PromptHistoryResult,
   ProviderModelCatalogUpdate,
@@ -170,7 +162,7 @@ interface DesktopWindowState {
 // added or renamed, the generated types shift, and a key present in GeneratedApp
 // but missing from AppBindings causes a type error here. Fix: add the new method
 // to AppBindings, then run `pnpm typecheck` to verify.
-export interface AppBindings {
+export interface AppBindings extends SessionCatalogBindings {
   Platform(): Promise<string>;
   MinimiseMainWindow(): Promise<void>;
   ToggleMaximiseMainWindow(): Promise<void>;
@@ -576,11 +568,6 @@ export interface AppBindings {
   CloseTerminalForTab(tabID: string, sessionID: string): Promise<void>;
   RenameTerminalForTab(tabID: string, sessionID: string, title: string): Promise<void>;
   ListProjectTree(): Promise<ProjectNode[]>;
-  GetProjectTreeSnapshot(): Promise<ProjectTreeSnapshot>;
-  ListProjectTopics(req: ProjectTopicPageRequest): Promise<ProjectTopicPage>;
-  GetTopicSummary(key: ProjectTopicKey): Promise<ProjectNode>;
-  GetSessionCatalogStatus(): Promise<SessionCatalogStatus>;
-  RebuildSessionCatalog(): Promise<void>;
   RenameProject(workspaceRoot: string, title: string): Promise<void>;
   SetProjectColor(workspaceRoot: string, color: string): Promise<void>;
   SetProjectPinned(workspaceRoot: string, pinned: boolean): Promise<void>;
@@ -894,21 +881,6 @@ export function onReady(cb: (tabId?: string) => void): () => void {
 export function onProjectTreeChanged(cb: () => void): () => void {
   if (realApp() && typeof window !== "undefined" && window.runtime) {
     return window.runtime.EventsOn("project-tree:changed", () => cb());
-  }
-  return () => {};
-}
-
-export function onProjectTreeChangedV2(cb: (event: ProjectTreeChangedV2) => void): () => void {
-  if (realApp() && typeof window !== "undefined" && window.runtime) {
-    return window.runtime.EventsOn("project-tree:changed-v2", (payload?: unknown) => {
-      if (!payload || typeof payload !== "object") return;
-      const event = payload as Partial<ProjectTreeChangedV2>;
-      cb({
-        revision: typeof event.revision === "number" ? event.revision : 0,
-        roots: asArray(event.roots).filter((root): root is string => typeof root === "string"),
-        reason: typeof event.reason === "string" ? event.reason : "changed",
-      });
-    });
   }
   return () => {};
 }
@@ -2430,6 +2402,7 @@ function makeMockApp(): AppBindings {
     }
   };
   return {
+    ...makeMockSessionCatalogBindings(cloneProjectTree),
     async MinimiseMainWindow() {
       console.info("mock MinimiseMainWindow");
     },
@@ -5173,35 +5146,6 @@ function makeMockApp(): AppBindings {
     async ListProjectTree() {
       return cloneProjectTree();
     },
-    async GetProjectTreeSnapshot() {
-      return {
-        revision: 1,
-        projects: cloneProjectTree().map((project) => ({ ...project, children: [] })),
-        catalog: { state: "ready", mode: "memory", revision: 1, indexed: 4, total: 4, repairPending: 0 },
-        indexed: 4,
-        total: 4,
-        indexingDone: true,
-      };
-    },
-    async ListProjectTopics(req: ProjectTopicPageRequest) {
-      const folder = req.scope === "global"
-        ? cloneProjectTree().find((item) => item.kind === "global_folder")
-        : cloneProjectTree().find((item) => item.kind === "project" && item.root === req.workspaceRoot);
-      const query = (req.query ?? "").trim().toLocaleLowerCase();
-      const all = asArray(folder?.children).filter((item) => !query || item.label.toLocaleLowerCase().includes(query));
-      const start = Math.max(0, Number.parseInt(req.cursor ?? "0", 10) || 0);
-      const limit = Math.min(200, Math.max(1, req.limit ?? 50));
-      const items = all.slice(start, start + limit);
-      return { items, nextCursor: start + items.length < all.length ? String(start + items.length) : undefined, revision: 1 };
-    },
-    async GetTopicSummary(key: ProjectTopicKey) {
-      const page = await this.ListProjectTopics({ scope: key.scope, workspaceRoot: key.workspaceRoot, limit: 200 });
-      return page.items.find((item) => item.topicId === key.topicId) ?? { key: "", kind: key.scope === "global" ? "global_topic" : "topic", label: "", children: [] };
-    },
-    async GetSessionCatalogStatus() {
-      return { state: "ready", mode: "memory", revision: 1, indexed: 4, total: 4, repairPending: 0 };
-    },
-    async RebuildSessionCatalog() {},
     async RenameProject(workspaceRoot: string, title: string) {
       const node = workspaceRoot
         ? mockProjectTree.find((item) => item.root === workspaceRoot)
